progress: 20.20% of code matched (2980/8216 functions, 3210/4540 objects)

+++ src/game/game_fn_800DB348.c
@@ -1,5 +1,5 @@
 extern int lbl_8064B788;
-extern int fn_801A98F4(unsigned short, int);
+extern int fn_801A98F4(int, int);
 
 void fn_800DB348(void)
 {
